pyproject.toml
[project]
name = "agneticrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=4.2.0,<5.0.0",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.128.5",
    "google-cloud-aiplatform>=1.136.0",
    "google-cloud-storage>=3.0.0",
    "google-generativeai>=0.8.6",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pydantic-ai-slim[google]>=1.56.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.9.0",
    "uvicorn>=0.40.0",
    "google-genai>=1.0.0",
    "aiosqlite>=0.22.1",
    "tabulate>=0.9.0",
    "tqdm>=4.67.3",
]
